Rename local to state
