+++ algebraicdb/src/executor/dbms/mod.rs
@@ -3,9 +3,9 @@ mod iter;
 use self::iter::*;
 use crate::ast::*;
 use crate::error_message::ErrorMessage;
-use crate::local::{DbState, DbmsState, ResourcesGuard};
 use crate::persistence::WriteToWal;
 use crate::pre_typechecker;
+use crate::state::{DbState, DbmsState, ResourcesGuard};
 use crate::table::{Cell, Schema, Table};
 use crate::typechecker;
 use crate::types::{Type, TypeId, TypeMap, Value};
