+++ algebraicdb/src/lib.rs
@@ -11,9 +11,9 @@ mod client;
 mod error_message;
 pub mod executor;
 pub mod grammar;
-pub mod local;
 mod persistence;
 mod pre_typechecker;
+pub mod state;
 pub mod table;
 mod typechecker;
 pub mod types;
